# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Tourtect API
  version: "1.0.0"
  description: |
    Tourtect backend API — community-powered travel safety and price transparency platform.
    All responses use snake_case JSON. Times are ISO 8601 UTC.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /v1/places:
    get:
      operationId: listPlaces
      tags: [places]
      summary: List places
      description: Returns a paginated list of places with optional text, region, and category filters.
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/LocaleParam"
        - $ref: "#/components/parameters/RequestIdParam"
        - name: q
          in: query
          description: Free-text search query
          schema:
            type: string
        - name: region_id
          in: query
          description: Filter by region identifier
          schema:
            type: string
        - name: category
          in: query
          description: Filter by place category
          schema:
            type: string
        - name: lat
          in: query
          description: Latitude for distance search
          schema:
            type: number
            format: double
        - name: lon
          in: query
          description: Longitude for distance search
          schema:
            type: number
            format: double
        - name: radius_m
          in: query
          description: Search radius in meters (requires lat/lon)
          schema:
            type: integer
            default: 5000
      responses:
        "200":
          description: Paginated list of place summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/places/{placeId}:
    get:
      operationId: getPlace
      tags: [places]
      summary: Get place details
      parameters:
        - $ref: "#/components/parameters/PlaceIdParam"
        - $ref: "#/components/parameters/LocaleParam"
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Place details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/feed:
    get:
      operationId: getFeed
      tags: [community]
      summary: Get an explainable organic community feed
      parameters:
        - { name: mode, in: query, required: true, schema: { type: string, enum: [following, nearby, latest, trending, safety] } }
        - { name: region_id, in: query, schema: { type: string } }
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200": { description: Explainable feed, content: { application/json: { schema: { $ref: "#/components/schemas/FeedResponse" } } } }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /v1/search:
    get:
      operationId: searchCommunity
      tags: [community]
      summary: Search public places and community content
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 2 } }
        - { name: tab, in: query, schema: { type: string, enum: [top, places, posts, price_reports, safety], default: top } }
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200": { description: Search results, content: { application/json: { schema: { $ref: "#/components/schemas/SearchResponse" } } } }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /v1/posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List published posts
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/LocaleParam"
        - $ref: "#/components/parameters/RequestIdParam"
        - name: place_id
          in: query
          description: Filter by linked place
          schema:
            type: string
            format: uuid
        - name: post_type
          in: query
          description: Filter by post type
          schema:
            $ref: "#/components/schemas/PostType"
      responses:
        "200":
          description: Paginated list of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/posts/drafts:
    post:
      operationId: createDraft
      tags: [posts]
      summary: Create a draft post
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDraftRequest"
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/posts/{postId}/publish:
    post:
      operationId: publishPost
      tags: [posts]
      summary: Publish a draft post
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/RequestIdParam"
      responses:
        "200":
          description: Post published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Post is not in draft state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/PostIdParam" }, { $ref: "#/components/parameters/LimitParam" }, { $ref: "#/components/parameters/RequestIdParam" } ]
      responses:
        "200": { description: Flat comment thread with parent identifiers, content: { application/json: { schema: { $ref: "#/components/schemas/CommentListResponse" } } } }
    post:
      operationId: createComment
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/PostIdParam" }, { $ref: "#/components/parameters/RequestIdParam" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateCommentRequest" } } } }
      responses:
        "201": { description: Comment created, content: { application/json: { schema: { $ref: "#/components/schemas/Comment" } } } }
  /v1/posts/{postId}/votes/useful:
    put:
      operationId: markPostUseful
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/PostIdParam" }, { $ref: "#/components/parameters/RequestIdParam" } ]
      responses: { "200": { description: Useful state set } }
    delete:
      operationId: unmarkPostUseful
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/PostIdParam" }, { $ref: "#/components/parameters/RequestIdParam" } ]
      responses: { "200": { description: Useful state cleared } }
  /v1/saved:
    get:
      operationId: listSavedPosts
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/RequestIdParam" } ]
      responses: { "200": { description: Saved public posts, content: { application/json: { schema: { $ref: "#/components/schemas/PostListResponse" } } } } }
  /v1/saved/posts/{postId}:
    put:
      operationId: savePost
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/PostIdParam" }, { $ref: "#/components/parameters/RequestIdParam" } ]
      responses: { "200": { description: Post saved } }
    delete:
      operationId: unsavePost
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/PostIdParam" }, { $ref: "#/components/parameters/RequestIdParam" } ]
      responses: { "200": { description: Post removed from saved list } }
  /v1/notifications:
    get:
      operationId: listNotifications
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/LimitParam" }, { $ref: "#/components/parameters/RequestIdParam" } ]
      responses: { "200": { description: Notifications, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationListResponse" } } } } }
    patch:
      operationId: updateNotifications
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/RequestIdParam" } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [notification_ids, read], properties: { notification_ids: { type: array, items: { type: string, format: uuid } }, read: { type: boolean } } } } } }
      responses: { "200": { description: Notification states updated } }
  /v1/follows:
    post:
      operationId: createFollow
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/RequestIdParam" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/FollowRequest" } } } }
      responses: { "200": { description: Follow created idempotently } }
    delete:
      operationId: deleteFollow
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/RequestIdParam" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/FollowRequest" } } } }
      responses: { "200": { description: Follow removed idempotently } }
  /v1/posts/{postId}/reports:
    post:
      operationId: reportPost
      tags: [community]
      parameters: [ { $ref: "#/components/parameters/PostIdParam" }, { $ref: "#/components/parameters/RequestIdParam" } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, enum: [safety, pii, harassment, spam, misinformation, other] }, details: { type: string } } } } } }
      responses: { "201": { description: Report queued for moderation } }
  /v1/blocks/{principalId}:
    parameters: [ { name: principalId, in: path, required: true, schema: { type: string, format: uuid } }, { $ref: "#/components/parameters/RequestIdParam" } ]
    put:
      operationId: blockPrincipal
      tags: [community]
      responses: { "200": { description: Principal blocked idempotently } }
    delete:
      operationId: unblockPrincipal
      tags: [community]
      responses: { "200": { description: Principal unblocked idempotently } }
  /v1/price-checks:
    post:
      operationId: createPriceCheck
      tags: [pricing]
      summary: Submit a price check
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceCheckRequest"
      responses:
        "200":
          description: Price insight result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceInsight"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/safety/assessments:
    post:
      operationId: createSafetyAssessment
      tags: [safety]
      summary: Submit a safety assessment request
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SafetyAssessmentRequest"
      responses:
        "200":
          description: Safety assessment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyAssessment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/assistant/sessions:
    post:
      operationId: createAssistantSession
      tags: [assistant]
      summary: Create an assistance session
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssistantSessionRequest"
      responses:
        "201":
          description: Versioned, short-lived assistance session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantSession"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          description: Session storage unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /v1/assistant/sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/AssistantSessionIdParam"
      - $ref: "#/components/parameters/RequestIdParam"
    get:
      operationId: getAssistantSession
      tags: [assistant]
      summary: Resume an owned assistance session
      responses:
        "200":
          description: Current redacted session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantSession"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAssistantSession
      tags: [assistant]
      summary: End and remove an assistance session
      responses:
        "204":
          description: Session state removed
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/assistant/sessions/{sessionId}/messages:
    post:
      operationId: createAssistantMessage
      tags: [assistant]
      summary: Send a typed input through the Tourtect Intelligence Layer
      parameters:
        - $ref: "#/components/parameters/AssistantSessionIdParam"
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMessageRequest"
      responses:
        "200":
          description: Evidence-backed structured assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Duplicate message identifier or stale session state
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/assistant/sessions/{sessionId}/confirmations:
    post:
      operationId: confirmAssistantAction
      tags: [assistant]
      summary: Confirm or reject a server-issued consequential action
      parameters:
        - $ref: "#/components/parameters/AssistantSessionIdParam"
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantConfirmationRequest"
      responses:
        "200":
          description: Confirmation outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantConfirmationResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Confirmation expired, already consumed, or state changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /v1/assistant/sessions/{sessionId}/feedback:
    post:
      operationId: createAssistantFeedback
      tags: [assistant]
      summary: Quarantine explicit assistant feedback or a user correction
      parameters:
        - $ref: "#/components/parameters/AssistantSessionIdParam"
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantFeedbackRequest"
      responses:
        "202":
          description: Feedback quarantined for validation or moderation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantFeedbackReceipt"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
components:
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Opaque cursor for pagination
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    LocaleParam:
      name: locale
      in: query
      description: Preferred locale for responses
      schema:
        $ref: "#/components/schemas/Locale"
    RequestIdParam:
      name: X-Request-ID
      in: header
      description: Client-provided request identifier for tracing
      schema:
        type: string
        format: uuid
    PlaceIdParam:
      name: placeId
      in: path
      required: true
      description: Place identifier
      schema:
        type: string
        format: uuid
    PostIdParam:
      name: postId
      in: path
      required: true
      description: Post identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    UnprocessableEntity:
      description: Request body validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
...
    PostType:
      type: string
      enum:
        - discussion
        - question
        - review
        - price_report
        - scam_report
        - tip
        - official_alert
        - external_link
...
    ProblemDetail:
      type: object
      required: [type, status, title]
      properties:
        type:
          type: string
          description: URI reference identifying the problem type
        status:
          type: integer
          description: HTTP status code
        title:
          type: string
          description: Short human-readable summary
        detail:
          type: string
          description: Human-readable explanation
        instance:
          type: string
          description: URI reference for this specific occurrence
        trace_id:
          type: string
          description: Request trace identifier
...
    PlaceDetail:
      type: object
      required: [place_id, name, category, coordinates, created_at]
      properties:
        place_id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        region_id:
          type: string
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        address:
          type: string
        description:
          type: string
        aliases:
          type: array
          items:
            type: string
        phone:
          type: string
        website:
          type: string
        opening_hours:
          type: string
        post_count:
          type: integer
        review_count:
          type: integer
        average_rating:
          type: number
          format: double
        freshness:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PlaceListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlaceSummary"
        pagination:
          $ref: "#/components/schemas/CursorInfo"
...
    CreateDraftRequest:
      type: object
      required: [post_type, original_locale, title, body]
      properties:
        post_type:
          $ref: "#/components/schemas/PostType"
        original_locale:
          $ref: "#/components/schemas/Locale"
        title:
          type: string
          minLength: 1
          maxLength: 500
        body:
          type: string
          minLength: 1
          maxLength: 50000
        region_id:
          type: string
        structured_data:
          type: object
          description: Type-specific draft fields validated and persisted by the community domain.
          additionalProperties: true
        place_ids:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 10
    PostListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        pagination:
          $ref: "#/components/schemas/CursorInfo"
...
    FeedResponse:
      type: object
      required: [items, pagination]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Post" } }
        pagination: { $ref: "#/components/schemas/CursorInfo" }
    SearchResponse:
      type: object
      required: [query, tab, places, posts]
      properties:
        query: { type: string }
        tab: { type: string }
        places: { type: array, items: { $ref: "#/components/schemas/PlaceSummary" } }
        posts: { type: array, items: { $ref: "#/components/schemas/Post" } }
...
    CreateCommentRequest:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 10000 }
        parent_comment_id: { type: string, format: uuid }
    CommentListResponse:
      type: object
      required: [items, pagination]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Comment" } }
        pagination: { $ref: "#/components/schemas/CursorInfo" }
...
    NotificationListResponse:
      type: object
      required: [items, pagination]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Notification" } }
        pagination: { $ref: "#/components/schemas/CursorInfo" }
...